requirements.txt
# MPRG - Multi-Path Reasoning Guard

pymongo
dnspython
voyageai
chromadb
flask
flask-cors
python-dotenv
sentence-transformers
numpy
requests
anthropic
